move csproj to fileparser
